// resources/js/api/generated/models/serverInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface ServerInfo {
  agent_cert_authority_fingerprint?: string;
  agent_cert_expires_at?: string;
  agent_cert_fingerprint?: string;
  agent_cert_issued_at?: string;
  backups_enabled: boolean;
  created_at: string;
  description: string;
  host: string;
  /** @minimum 0 */
  id: number;
  is_active: boolean;
  name: string;
  port: number;
  skip_ssl_verification: boolean;
  updated_at: string;
}